+++ src/dkr/core/__init__.py
@@ -1,1 +1,1 @@
-from . import ends
+from . import artifacting, didding, ends, habs, resolving, webbing
